Домашнее задание №6 (часть 1)

+++ Assets/Scripts/Enemy/EnemyAsset.cs
@@ -6,7 +6,7 @@ namespace Enemy
     public class EnemyAsset : ScriptableObject
 
     {
-        public int StartHealth;
+        public float StartHealth;
 
         public bool IsFlyingEnemy;
